# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Scandere AI Store API",
        "description": "REST API for digital products e-commerce platform. Sell and manage digital documents (PDFs, Excel, Word) with secure downloads, payments via Stripe, and comprehensive admin panel.",
        "contact": {
            "name": "Scandere Support",
            "email": "team@scandere.info"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.scandereai.store",
            "description": "Production API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/admin/messages": {
            "get": {
                "tags": [
                    "Admin - Messages"
                ],
                "summary": "List all contact messages (admin)",
                "description": "Get paginated list of all contact form submissions. Requires admin authentication.",
                "operationId": "5e382b38c49c3ded62fe315c5f237cb9",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated messages list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "first_name": {
                                                        "type": "string"
                                                    },
                                                    "last_name": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "message": {
                                                        "type": "string"
                                                    },
                                                    "is_read": {
                                                        "type": "boolean"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 30
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden (not admin)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/messages/{id}/read": {
            "patch": {
                "tags": [
                    "Admin - Messages"
                ],
                "summary": "Mark message as read (admin)",
                "description": "Mark a contact message as read. Requires admin authentication.",
                "operationId": "b2ceb4b131b5c9ccbbf969d61913b6c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Message ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message marked as read",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Marked as read."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden (not admin)"
                    },
                    "404": {
                        "description": "Message not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/stats": {
            "get": {
                "tags": [
                    "Admin - Dashboard"
                ],
                "summary": "Get admin dashboard statistics",
                "description": "Get comprehensive statistics for admin dashboard including users, subscribers, products, orders, revenue, and recent activity. Requires admin authentication.",
                "operationId": "deeeec8019a30ad6f7ea8c7327d49480",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_users": {
                                            "description": "Total registered users",
                                            "type": "integer",
                                            "example": 150
                                        },
                                        "total_subscribers": {
                                            "description": "Total active newsletter subscribers",
                                            "type": "integer",
                                            "example": 320
                                        },
                                        "total_products": {
                                            "description": "Total products",
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "total_orders": {
                                            "description": "Total completed orders",
                                            "type": "integer",
                                            "example": 85
                                        },
                                        "total_revenue": {
                                            "description": "Total revenue from completed orders",
                                            "type": "number",
                                            "format": "float",
                                            "example": 2123.15
                                        },
                                        "unread_messages": {
                                            "description": "Number of unread contact messages",
                                            "type": "integer",
                                            "example": 7
                                        },
                                        "recent_orders": {
                                            "description": "10 most recent completed orders",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "total": {
                                                        "type": "number"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "completed"
                                                    },
                                                    "paid_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "first_name": {
                                                                "type": "string"
                                                            },
                                                            "last_name": {
                                                                "type": "string"
                                                            },
                                                            "email": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "recent_subscribers": {
                                            "description": "10 most recent subscribers",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "first_name": {
                                                        "type": "string"
                                                    },
                                                    "source": {
                                                        "type": "string",
                                                        "example": "newsletter"
                                                    },
                                                    "subscribed_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden (not admin)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format: Bearer {token}"
            }
        }
    },
...
}